+++ Source/Relativity.Sync/Storage/IBatch.cs
@@ -12,9 +12,9 @@ namespace Relativity.Sync.Storage
 		BatchStatus Status { get; }
 		int TransferredItemsCount { get; }
 		int TaggedItemsCount { get; }
-		int MetadataBytesTransferred { get; }
-		int FilesBytesTransferred { get; }
-		int TotalBytesTransferred { get; }
+		long MetadataBytesTransferred { get; }
+		long FilesBytesTransferred { get; }
+		long TotalBytesTransferred { get; }
 		int TotalItemsCount { get; }
 
 		Task SetFailedItemsCountAsync(int failedItemsCount);
@@ -23,9 +23,9 @@ namespace Relativity.Sync.Storage
 		Task SetStatusAsync(BatchStatus status);
 		Task SetTransferredItemsCountAsync(int transferredItemsCount);
 		Task SetTaggedItemsCountAsync(int taggedItemsCount);
-		Task SetMetadataBytesTransferredAsync(int metadataBytesTransferred);
-		Task SetFilesBytesTransferredAsync(int filesBytesTransferred);
-		Task SetTotalBytesTransferredAsync(int totalBytesTransferred);
+		Task SetMetadataBytesTransferredAsync(long metadataBytesTransferred);
+		Task SetFilesBytesTransferredAsync(long filesBytesTransferred);
+		Task SetTotalBytesTransferredAsync(long totalBytesTransferred);
 		Task SetStartingIndexAsync(int newStartIndex);
 	}
 }
